backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
itsdangerous==2.1.2
pydantic>=2.6
google-generativeai==0.3.2
Pillow>=10.0.0
